Move subscription struct to its own file
